+++ bank-feeds/src/codatbankfeeds/sdk.py
@@ -68,7 +68,9 @@ class CodatBankFeeds:
         if client is None:
             client = requests_http.Session()
         
+        
         security_client = utils.configure_security_client(client, security)
+        
         
         if server_url is not None:
             if url_params is not None: