+++ bank-feeds/src/codatbankfeeds/sdkconfiguration.py
@@ -20,8 +20,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '3.0.0'
-    sdk_version: str = '3.1.0'
-    gen_version: str = '2.116.0'
+    sdk_version: str = '3.2.0'
+    gen_version: str = '2.155.1'
+    user_agent: str = 'speakeasy-sdk/python 3.2.0 2.155.1 3.0.0 codat-bankfeeds'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> tuple[str, dict[str, str]]: